pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "mm-cxr-diag"
version = "0.2.0"
description = "Multi-stage multimodal classifier for NIH ChestX-ray14 — abnormality gate + pathology classifier"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Travis Nesbit" }]
keywords = ["chest-xray", "multimodal", "deep-learning", "medical-imaging", "pytorch"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: MIT License",
  "Operating System :: POSIX :: Linux",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

# Torch is NOT pinned to a CUDA build here. Install torch first from the
# appropriate PyTorch index (scripts/install_torch.sh, or the Dockerfile),
# then install this package.
dependencies = [
  "torch>=2.3,<3",
  "torchvision>=0.18,<1",
  "numpy>=1.26",
  "pandas>=2.0",
  "pillow>=10.0",
  "scikit-learn>=1.3",
  "tqdm>=4.65",
  "matplotlib>=3.7",
  "seaborn>=0.12",
  "pyyaml>=6.0",
  "python-dotenv>=1.0",
  "tabulate>=0.9",
  "timm>=1.0",
  "kagglehub>=0.2",
  "pydantic>=2.5",
  "pydantic-settings>=2.1",
  "albumentations>=1.4",
  "typer>=0.12",
]

[project.optional-dependencies]
serve = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "python-multipart>=0.0.9",
]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.4",
  "black>=24.0",
  "mypy>=1.9",
  "pre-commit>=3.6",
]
viz = [
  "grad-cam>=1.4",
  "shap>=0.44",
]
all = [
  "mm-cxr-diag[serve,dev,viz]",
]

[project.scripts]
mm-cxr-diag = "mm_cxr_diag.cli.main:app"

[project.urls]
Repository = "https://github.com/geekmdtravis/mm-cxr-diag"

[tool.hatch.build.targets.wheel]
packages = ["mm_cxr_diag"]

[tool.hatch.build.targets.sdist]
include = [
  "mm_cxr_diag/",
  "configs/",
  "scripts/",
  "tests/",
  "README.md",
  "LICENSE",
  "pyproject.toml",
]

# --- Tooling -------------------------------------------------------------

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = [
  "-ra",
  "--strict-markers",
  "--strict-config",
]
markers = [
  "slow: marks tests that take more than ~5s (run with -m slow)",
  "real_data: requires MM_CXR_NIH_DIR env var pointing at the NIH dataset",
]

[tool.coverage.run]
source = ["mm_cxr_diag"]
branch = true

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "if __name__ == .__main__.:",
  "raise NotImplementedError",
]

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["mm_cxr_diag"]

[tool.ruff.lint]
select = [
  "E", "W",  # pycodestyle
  "F",        # pyflakes
  "I",        # isort
  "B",        # flake8-bugbear
  "UP",       # pyupgrade
  "SIM",      # flake8-simplify
]
ignore = [
  "E203",  # whitespace before ':' — black-compatible
  "E501",  # line too long — black handles it
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]
"tests/**" = ["B011"]
# FastAPI's Depends/File/Form/Body are idiomatic in parameter defaults.
"mm_cxr_diag/service/routes.py" = ["B008"]

[tool.black]
line-length = 88
target-version = ["py311", "py312"]

[tool.mypy]
python_version = "3.11"
# Scoped mypy: only the parts of the codebase where strong typing pays off.
# Training/model code is skipped — tensor-shape typing is noise.
files = ["mm_cxr_diag/config", "mm_cxr_diag/cli", "mm_cxr_diag/service", "mm_cxr_diag/inference"]
strict_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
ignore_missing_imports = true
